+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## owncloud-onlyoffice (1.0.1)
 - fix exception when versions app is disabled
 - adding protocol to document server url
-
+- onlyofficeOpen is default action
 
 ## owncloud-onlyoffice (1.0.0)
 - Initial release